requirements.txt
Flask==2.3.3
web3==6.11.0
opencv-python-headless==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
python-dotenv==1.0.0
eth-account==0.9.0
PyYAML==6.0.1
redis==5.0.1
asyncio==3.4.3
flask-limiter==3.5.0
gunicorn==21.2.0